autouse env fix
