+++ cstar/tests/integration_tests/test_cstar_test_blueprints.py
@@ -55,15 +55,9 @@ class TestCStar:
         log (logging.Logger):
             Logger instance for logging messages during test execution.
         """
-        dotenv_path = tmp_path / ".cstar.env"
         ext_root = tmp_path / "externals"
 
         with (
-            mock.patch(
-                "cstar.system.environment.CStarEnvironment.user_env_path",
-                new_callable=mock.PropertyMock,
-                return_value=dotenv_path,
-            ),
             mock.patch(
                 "cstar.system.environment.CStarEnvironment.lmod_path", mock_lmod_path
             ),
