+++ cstar/tests/unit_tests/conftest.py
@@ -3,11 +3,9 @@ import logging
 from collections.abc import Callable, Generator
 from pathlib import Path
 from unittest import mock
-from unittest.mock import PropertyMock, patch
 
 import dotenv
 import pytest
-from _pytest.tmpdir import TempPathFactory
 
 from cstar.base import AdditionalCode, Discretization, ExternalCodeBase, InputDataset
 from cstar.base.datasource import DataSource
@@ -309,21 +307,9 @@ def system_dotenv_dir(tmp_path: Path) -> Path:
 
 
 @pytest.fixture
-def mock_system_name() -> str:
+def mock_system_name(scope="session") -> str:
     # A name for the mock system/platform executing the tests.
     return "mock_system"
-
-
-@pytest.fixture(scope="session")
-def mock_user_env_name() -> str:
-    """Return a unique name for a temporary user .env config file.
-
-    Returns
-    -------
-    str
-        The name of the .env file
-    """
-    return ".mock.env"
 
 
 @pytest.fixture
@@ -424,17 +410,7 @@ def custom_user_env(
     return functools.partial(_write_custom_env, dotenv_path)
 
 
-@pytest.fixture(scope="session", autouse=True)
-def dont_touch_user_env(tmp_path_factory: TempPathFactory, mock_user_env_name: str):
-    with patch(
-        "cstar.system.environment.CStarEnvironment.user_env_path",
-        new_callable=PropertyMock,
-        return_value=tmp_path_factory.mktemp("user_env") / mock_user_env_name,
-    ):
-        yield
-
-
-@pytest.fixture
+@pytest.fixture(scope="session")
 def mock_lmod_filename() -> str:
     """Return a complete path to an empty, temporary .lmod config file for tests.
