+++ cstar/tests/unit_tests/system/test_environment.py
@@ -159,11 +159,6 @@ class TestSetupEnvironmentFromFiles:
         )  # fmt: skip
         # Patch the root path and expanduser to point to our temporary files
         with (
-            patch(
-                "cstar.system.environment.CStarEnvironment.user_env_path",
-                new_callable=PropertyMock,
-                return_value=dotenv_path,
-            ),
             patch.object(
                 cstar.system.environment.CStarEnvironment, "package_root", new=tmp_path
             ),
@@ -232,11 +227,6 @@ class TestSetupEnvironmentFromFiles:
                 "cstar.system.environment.CStarEnvironment.system_env_path",
                 new_callable=PropertyMock,
                 return_value=system_dotenv_path,
-            ),
-            patch(
-                "cstar.system.environment.CStarEnvironment.user_env_path",
-                new_callable=PropertyMock,
-                return_value=dotenv_path,
             ),
         ):
             # Instantiate the environment to trigger loading the environment variables
@@ -432,11 +422,6 @@ class TestSetupEnvironmentFromFiles:
                 new_callable=Mock,
                 side_effect=[env0, env1],
             ) as loader,
-            patch(
-                "cstar.system.environment.CStarEnvironment.user_env_path",
-                new_callable=PropertyMock,
-                return_value=dotenv_path,
-            ),
         ):
             # Instantiate the environment to trigger loading the environment variables
             env = CStarEnvironment(
